+++ tests/solver/_strategies/_initialization/test_init_dummy.py
@@ -1,0 +1,19 @@
+import pytest
+
+from max_div.solver._strategies import InitializationStrategy
+
+from ._helpers import new_solver_state
+
+
+@pytest.mark.parametrize("problem_has_constraints", [True, False])
+def test_init_dummy(problem_has_constraints: bool):
+    # --- arrange -----------------------------------------
+    solver_state = new_solver_state(problem_has_constraints)
+    strategy = InitializationStrategy.dummy()
+
+    # --- act ---------------------------------------------
+    strategy.initialize(solver_state)
+    score = solver_state.score
+
+    # --- assert ------------------------------------------
+    assert score.size == 1.0, "Selection size should be equal to k after initialization"